Degeneracy testing for extreme configurations
Enforcing triangle areas to be non-close to zero on all views.
This may make the solver fail more often on tests for perfect sythetic
data. But in practice we may avoid spending expensive iterations on
bad configurations that are hit by chance in RANSAC, just to realize in
inlier computation they are bad. The solver may actually work a lot
better than expected for these seemingly degenerate configurations.

We may want to increase these precision value to get the app to remove
more unstable views, but that requires experimentation on total time.